add ESILimiter to watch ESI error headers
Watch the ESI error limit response headers. Limit the number of
simultaneous outstanding requests to be equal to the ESI error limit. If
the error limit is zero, then we will prevent all outgoing requests
until the timeout occurs, at which point we will allow one request, in
order to learn what the new limit is.