make refresh token logic better handle cancels
The old logic appeared to be vulnerable to a variety of issues around
task cancellation, especially that the refresh task getting cancelled
resulted in routing the possibly-successful refresh token refresh result
to nowhere.

Move the entire session refresh token update logic to a routine that we
will treat as atomic, and protect that against cancellation instead.