fix perpetual bad-state session bug
Instead of perpetually caching results both positive and negative, only
cache positive results. Since we were previously caching the task
result instead of the positive results only, it was possible to create a
session that was permanently stuck in the `CharacterNeedsUpdated` state.

Purge cached character sessions if they go unused for five minutes.

When a character goes through the authorization flow, if there is a
pre-existing session for that character, set its access token at that
time. I am not sure if this is the best solution, but it solves a
problem where this wasn't occurring when it was supposed to be.

Remove the `valid` flag from the `Character` namedtuple as it is only
relevant for one UI endpoint, and having a `valid` flag on a `Character`
attached to a `Session` makes one wary, as otherwise the character's
validity state would need to change based on the contents of other
attributes.