fix CharacterHover bad behavior on screen right
If the scrollbar of characters was scrolled far to the right, the
CharacterHover element would be drawn off-screen. Even without this, the
element would crowd up against the right-side edge for characters on the
far right side of the page.

Fix this by calculating the screen position of the character DOM being
hovered and then position the CharacterHover element as an offset from
either the left or right side of the page. This looks much nicer and has
correct behavior if the scrollbar is scrolled to the right.